renamed functions and moved files
