+++ src/api/PersonAPI.ts
@@ -59,6 +59,36 @@ export const deletePerson = async (id: number) => {
 
 export const getProjectOfPersonByID = async (id: number) => {
     try {
+        const route = "project";
+        const res = await api.get(`${personRoute}/${route}/${id}`);
+        const project = res.data.payload as Project; 
+        if (!project) {
+            return;
+        }
+        return project;
+    } catch (err) {
+        console.error("Error fetching project:", err);
+        throw err;
+    }
+};
+
+export const getPositionOfPersonByID = async (id: number) => {
+    try {
+        const route = "position";
+        const res = await api.get(`${personRoute}/${route}/${id}`);
+        const position = res.data.payload as Position;
+        if (!position) {
+            return;
+        }
+        return position;
+    } catch (err) {
+        console.error("Error fetching position:", err);
+        throw err;
+    }
+};
+
+export const getProjectOfPersonByIDResource = async (id: number) => {
+    try {
         const route = "/projects";
         const res = await api.get(`${route}/${id}`);
         const project = res.data.payload as Project; 
@@ -72,7 +102,7 @@ export const getProjectOfPersonByID = async (id: number) => {
     }
 };
 
-export const getPositionOfPersonByID = async (id: number) => {
+export const getPositionOfPersonByIDResource = async (id: number) => {
     try {
         const route = "/positions";
         const res = await api.get(`${route}/${id}`);
